Finish A-CodingStandard
