+++ src/main/java/duke/commands/Command.java
@@ -1,9 +1,9 @@
-package duke.Commands;
+package duke.commands;
 
-import duke.Exceptions.DukeException;
-import duke.Storage.Storage;
-import duke.TaskList.TaskList;
-import duke.Ui.Ui;
+import duke.exceptions.DukeException;
+import duke.storage.Storage;
+import duke.tasklist.TaskList;
+import duke.ui.Ui;
 
 public interface Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
