+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,10 +1,10 @@
-package duke.Commands;
+package duke.commands;
 
-import duke.Exceptions.DukeException;
-import duke.Exceptions.TaskException;
-import duke.Storage.Storage;
-import duke.TaskList.TaskList;
-import duke.Ui.Ui;
+import duke.exceptions.DukeException;
+import duke.exceptions.TaskException;
+import duke.storage.Storage;
+import duke.tasklist.TaskList;
+import duke.ui.Ui;
 
 public class DeleteCommand implements Command {
 
@@ -18,8 +18,9 @@ public class DeleteCommand implements Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         try {
             int taskNum = Integer.parseInt(fullCommand.substring(7));
-            if (taskNum > tasks.size() || taskNum < 1)
+            if (taskNum > tasks.size() || taskNum < 1) {
                 throw new TaskException("Please enter a valid task number!");
+            }
             tasks.deleteTask(taskNum, storage, ui);
         } catch (NumberFormatException | StringIndexOutOfBoundsException e) {
             throw new TaskException("Please enter a valid task number!");
