+++ src/main/java/duke/commands/ListCommand.java
@@ -1,18 +1,19 @@
-package duke.Commands;
+package duke.commands;
 
-import duke.Exceptions.DukeException;
-import duke.Storage.Storage;
-import duke.TaskList.TaskList;
-import duke.Ui.Ui;
+import duke.exceptions.DukeException;
+import duke.storage.Storage;
+import duke.tasklist.TaskList;
+import duke.ui.Ui;
 
 public class ListCommand implements Command {
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        if (tasks.size() == 0)
+        if (tasks.size() == 0) {
             ui.printNoTasks();
-        else
+        } else {
             ui.printTasks(tasks);
+        }
     }
 
     @Override
