+++ src/main/java/duke/exceptions/InvalidInputException.java
@@ -1,4 +1,4 @@
-package duke.Exceptions;
+package duke.exceptions;
 
 public class InvalidInputException extends DukeException {
     public InvalidInputException(String message) {
