+++ src/main/java/duke/exceptions/ListException.java
@@ -1,4 +1,4 @@
-package duke.Exceptions;
+package duke.exceptions;
 
 public class ListException extends DukeException {
     public ListException(String message) {
