+++ src/main/java/duke/exceptions/TaskException.java
@@ -1,4 +1,4 @@
-package duke.Exceptions;
+package duke.exceptions;
 
 public class TaskException extends DukeException {
     public TaskException(String message) {
