+++ src/main/java/duke/task/Task.java
@@ -1,4 +1,4 @@
-package duke.Task;
+package duke.task;
 
 public class Task {
     private String description;
