+++ src/main/java/duke/ui/Ui.java
@@ -1,10 +1,10 @@
-package duke.Ui;
+package duke.ui;
 
 import java.util.ArrayList;
 import java.util.Scanner;
 
-import duke.Task.Task;
-import duke.TaskList.TaskList;
+import duke.task.Task;
+import duke.tasklist.TaskList;
 
 public class Ui {
 
@@ -29,10 +29,11 @@ public class Ui {
     }
 
     public void showSuccessfulLoad(TaskList tasks) {
-        if (tasks.size() == 0)
+        if (tasks.size() == 0) {
             System.out.println("  Data file was loaded but no tasks could be found.");
-        else
+        } else {
             System.out.println("  Tasks loaded successfully!");
+        }
     }
 
     public void showError(String message) {
@@ -53,20 +54,22 @@ public class Ui {
     }
 
     public void printTask(Task task, ArrayList<Task> taskList, boolean toAdd) {
-        if (toAdd)
+        if (toAdd) {
             System.out.println("  Got it. I've added this task:");
-        else
+        } else {
             System.out.println("  Noted. I've removed this task:");
+        }
         System.out.println(String.format("  %s", task.toString()));
         System.out.println(String.format("  Now you have %d %s in the list.", taskList.size(),
                 taskList.size() == 1 ? "task" : "tasks"));
     }
 
     public void printMarkTask(Task task, boolean toMark) {
-        if (toMark)
+        if (toMark) {
             System.out.println("  Nice! I've marked this task as done:");
-        else
+        } else {
             System.out.println("  I've marked this task as not done:");
+        }
         System.out.println(String.format("  %s", task.toString()));
     }
