+++ src/test/java/duke/task/DeadlineTest.java
@@ -1,4 +1,4 @@
-package duke.Task;
+package duke.task;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
