+++ src/test/java/duke/task/TodoTest.java
@@ -1,4 +1,4 @@
-package duke.Task;
+package duke.task;
 
 import org.junit.jupiter.api.Test;
